Remove extraction Pydantic models
